python/src/requirements.txt
ariadne==0.17.0
flask
pg8000
python-dotenv